requirements.txt
setuptools
ai2thor==5.0.0
blinker==1.7.0
botocore==1.34.79
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
idna==3.6
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.26.4
pillow==10.3.0
progressbar2==4.4.2
python-dateutil==2.9.0.post0
python-utils==3.8.2
python-xlib==0.33
PyYAML==6.0.1
requests==2.31.0
typing-extensions==4.11.0
urllib3==1.26.18
werkzeug==3.0.2
zipp==3.18.1
python-dotenv
demjson3
langchain==0.2.14
langchain_community
openai
